Provision/Post-Provision/requirements.txt
# Requirements for Azure Functions
azure-functions
redis
azure-eventgrid
